+++ doc/changelog.md
@@ -2,6 +2,9 @@
 
 <h3>New features</h3>
 
+* `qjit` adheres to user-specified `mcm_method` given to the `QNode`.
+  [(#798)](https://github.com/PennyLaneAI/catalyst/pull/798)
+
 * The `dynamic_one_shot` transform uses a single auxiliary tape which is repeatedly simulated `n_shots` times to simulate hardware-like results.
   The loop over shots is executed with `catalyst.vmap`.
   [(#5617)](https://github.com/PennyLaneAI/pennylane/pull/5617)
@@ -87,6 +90,29 @@
     return excluded_module.submodule.func(x)
 
   ```
+
+* Add support for accelerating classical processing via JAX with `catalyst.accelerate`.
+  [(#805)](https://github.com/PennyLaneAI/catalyst/pull/805)
+
+  Classical code that can be just-in-time compiled with JAX can now be seamlessly just
+  in time compiled with `catalyst.accelerate` and included within QJIT-compiled functions.
+  `catalyst.accelerate` can be used as a
+  decorator without specifying a device:
+
+  ```python
+  @accelerate(dev=jax.devices("gpu")[0])
+  def classical_fn(x):
+      return jnp.sin(x) ** 2
+
+  @qjit
+  def hybrid_fn(x):
+      y = classical_fn(jnp.sqrt(x)) # will be executed on a GPU
+      return jnp.cos(y)
+  ```
+
+  Available devices can be retrieved via
+  `jax.devices()`. If not provided, the default value of
+  `jax.devices()[0]` as determined by JAX will be used.
 
 <h3>Improvements</h3>
 
@@ -309,6 +335,7 @@ Haochen Paul Wang,
 Lee James O'Riordan,
 Mehrdad Malekmohammadi,
 Vincent Michaud-Rioux,
+Mudit Pandey,
 Raul Torres,
 Sergei Mironov.